Fix issue when no ttyS0
Closes #15

The issue appears to be that the Firecracker provided `init` binary
changes I/O privileges to enable I/O-port read/write on port 0x3f0 for
all processes (likely by calling `iopl`). This _seems_ to be conditional
on the existence of `ttyS0` somehow (we don't have the source, so not
sure---maybe the existence of ttyS0 is used to signal debug mode or
something).

As a result, all of our calls to `outl` were seg-faulting (which is the
result of calling `outl()` with the wrong permissions).

Instead, we revert to OpenRC's normal `init` binary and explicitly
enable I/O privileges in our `outl` binary.