A runtime for programs with no runtime
Includes a C-program to act as the "runtime", in charge of the
snapshotting, mounting the app file system, openning and parsing the
request/response console device.

The interface is that each function invocation is a new instance of
`/srv/workload`, which can be any Linux executable, but likely a shell
script. The request JSON is passed as the first command-line argument,
and the response is taken as a line-terminated JSON from the standard
out of the process.